Retrieve Epitope Residue Sites for Influenza HA Subtypes

This function returns epitope residue positions (sites A–E and optionally all sites combined) for a given influenza HA subtype. It supports H1N1, H3N2, and influenza B lineages (Victoria, Yamagata, or pre-split) with options for harmonizing B lineages and adjusting for signal peptide inclusion.

Usage

get_pepitope_sites(
  subtype,
  sites = c("a", "b", "c", "d", "e"),
  harmonize_b_lineages = TRUE,
  seq_includes_signal_peptide = TRUE
)

Arguments

subtype

Character. HA subtype or lineage, e.g., `"H1N1"`, `"A(H3N2)"`, `"B/Victoria"`, etc.

sites

Character vector. Which epitope sites to return. Must be a subset of `c("a", "b", "c", "d", "e", "all_epitopes")`. Defaults to `c("a", "b", "c", "d", "e")`.

harmonize_b_lineages

Logical. If `TRUE`, merges all B lineages into a unified "B" category. Defaults to `TRUE`.

seq_includes_signal_peptide

Logical. If `TRUE`, epitope sites will be offset by the length of the HA signal peptide. Defaults to `TRUE`.

Value

A named list of integer vectors, with names corresponding to the requested `sites`.

Examples

get_pepitope_sites("A(H1N1)", sites = c("a", "b"))
#> $a
#>  [1] 135 137 138 139 143 144 145 146 149 150 151 152 154 156 157 158 159 160 163
#> [20] 164 166 182 269 270
#> 
#> $b
#>  [1] 141 142 169 170 171 172 173 174 177 179 200 201 202 203 204 206 207 208 210
#> [20] 211 212 213
#> 
get_pepitope_sites("B/Yamagata", harmonize_b_lineages = FALSE)
#> $a
#>  [1] 135 136 137 139 140 148 149 150 151 153 155 156 158 160 161 162 163 164 165
#> [20] 169 171 190
#> 
#> $b
#>  [1] 141 143 147 174 175 176 177 178 179 181 185 187 209 210 211 212 213 215 216
#> [20] 217 219 220 221 222
#> 
#> $c
#>  [1]  48  49  50  51  52  53  54 302 304 305 306 307 322 328 330 331 333 334 336
#> [20] 337 338 339 340
#> 
#> $d
#>  [1] 107 115 116 130 134 189 192 193 195 196 197 198 199 200 201 203 225 227 231
#> [20] 232 233 236 237 238 239 240 241 242 243 246 255 256 257 258 259 267 268 269
#> [39] 270 271
#> 
#> $e
#>  [1]  56  58  62  70  72  73  77  85  87  89  91  92  93  94  97  98  99 102 103
#> [20] 105 122 286 289 290 293
#>